+++ app/controllers/posts_controller.rb
@@ -1,60 +1,67 @@
 class PostsController < ApplicationController
   def index
     matching_posts = Post.all
-
     @list_of_posts = matching_posts.order({ :created_at => :desc })
-
     render({ :template => "posts/index" })
   end
 
   def show
     the_id = params.fetch("path_id")
-
-    matching_posts = Post.where({ :id => the_id })
-
-    @the_post = matching_posts.at(0)
-
-    render({ :template => "posts/show" })
+    @the_post = Post.find_by(id: the_id)
+    
+    if @the_post.nil?
+      redirect_to("/posts", { :alert => "Post not found." })
+    else
+      render({ :template => "posts/show" })
+    end
   end
 
   def create
     the_post = Post.new
-    the_post.title = params.fetch("query_title")
-    the_post.body = params.fetch("query_body")
-    the_post.expires_on = params.fetch("query_expires_on")
-    the_post.board_id = params.fetch("query_board_id")
-
+    the_post.title = params.fetch("title")
+    the_post.body = params.fetch("body")
+    the_post.expires_on = params.fetch("expires_on")
+    the_post.board_id = params.fetch("board_id")
+    
     if the_post.valid?
       the_post.save
-      redirect_to("/posts", { :notice => "Post created successfully." })
+      redirect_to("/boards/#{the_post.board_id}", { :notice => "Post created successfully." })
     else
-      redirect_to("/posts", { :alert => the_post.errors.full_messages.to_sentence })
+      redirect_to("/boards/#{the_post.board_id}", { :alert => the_post.errors.full_messages.to_sentence })
     end
   end
 
   def update
     the_id = params.fetch("path_id")
-    the_post = Post.where({ :id => the_id }).at(0)
-
-    the_post.title = params.fetch("query_title")
-    the_post.body = params.fetch("query_body")
-    the_post.expires_on = params.fetch("query_expires_on")
-    the_post.board_id = params.fetch("query_board_id")
-
-    if the_post.valid?
-      the_post.save
-      redirect_to("/posts/#{the_post.id}", { :notice => "Post updated successfully."} )
+    the_post = Post.find_by(id: the_id)
+    
+    if the_post.nil?
+      redirect_to("/posts", { :alert => "Post not found." })
     else
-      redirect_to("/posts/#{the_post.id}", { :alert => the_post.errors.full_messages.to_sentence })
+      the_post.title = params.fetch("query_title")
+      the_post.body = params.fetch("query_body")
+      the_post.expires_on = params.fetch("query_expires_on")
+      the_post.board_id = params.fetch("query_board_id")
+      
+      if the_post.valid?
+        the_post.save
+        redirect_to("/posts/#{the_post.id}", { :notice => "Post updated successfully."} )
+      else
+        redirect_to("/posts/#{the_post.id}", { :alert => the_post.errors.full_messages.to_sentence })
+      end
     end
   end
 
   def destroy
     the_id = params.fetch("path_id")
-    the_post = Post.where({ :id => the_id }).at(0)
-
-    the_post.destroy
-
-    redirect_to("/posts", { :notice => "Post deleted successfully."} )
+    the_post = Post.find_by(id: the_id)
+    
+    if the_post.nil?
+      redirect_to("/posts", { :alert => "Post not found." })
+    else
+      board_id = the_post.board_id
+      the_post.destroy
+      redirect_to("/boards/#{board_id}", { :notice => "Post deleted successfully."} )
+    end
   end
 end
